refactor: make faucet pkg public
for programmatic usage.

solves a part of #tendermint/starport/570.
