+++ faucet/cmd.go
