+++ faucet/key.go
