+++ faucet/options.go
